+++ src/Tests/Dfe.Complete.CypressTests/cypress/api/taskHelper.ts
@@ -135,7 +135,7 @@ export class TaskHelper {
             notApplicable: false,
             received: false,
             cleared: false,
-            sent: false,
+            draftSaved: false,
             saved: false,
             signed: false,
             signedSecretaryState: false,
@@ -161,7 +161,7 @@ export class TaskHelper {
                     notApplicable: false,
                     received: true,
                     cleared: true,
-                    sent: true,
+                    draftSaved: true,
                     saved: true,
                     signed: true,
                     signedSecretaryState: true,
@@ -377,9 +377,9 @@ export class TaskHelper {
             projectType: projectType,
             received: false,
             cleared: false,
-            sent: false,
             signed: false,
             saved: false,
+            draftSaved: false,
             signedSecretaryState: false,
         };
         switch (status) {
@@ -395,9 +395,9 @@ export class TaskHelper {
                     projectType: projectType,
                     received: true,
                     cleared: true,
-                    sent: true,
                     signed: true,
                     saved: true,
+                    draftSaved: true,
                     signedSecretaryState: true,
                 });